% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod_municipios_17}
\alias{cod_municipios_17}
\title{cod_municipios_17.rda}
\format{A data frame with 8124 rows and 4 variables: Cerdedo-Cotobade se unieron en 2016
\itemize{
\item CODAUTO: Código de CCAA
\item CPRO: Código de la provincia
\item CMUN: Código del municipio
\item DC: Código de control
\item NOMBRE: Nombre del municipio
}}
\source{
\url{http://www.ine.es}
}
\usage{
cod_municipios_17
}
\description{
Codigos del INE para municipios españoles a 1 enero 2017
}
\keyword{datasets}
