% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod_provincias}
\alias{cod_provincias}
\title{This file documents datasets}
\format{A data frame with 52 rows and 4 variables:
\itemize{
\item CODAUTO: Código de CCAA
\item CCAA: Nombre de la CCAA
\item CPRO: Código de la provincia
\item PROVINCIA: Nombre de la provincia
}}
\source{
\url{http://www.ine.es}
}
\usage{
cod_provincias
}
\description{
cod_provincias.rda
}
\details{
Codigos del INE para provincias y CCAA españolas
}
\keyword{datasets}
