% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjppkgRdata01-package.R
\docType{package}
\name{pjppkgRdata01-package}
\alias{pjppkgRdata01}
\alias{pjppkgRdata01-package}
\title{pjppkgRdata01: Datasets to use in my analysis and blog}
\description{
Provides datasets for easy loading in R, mainly in my blog
R & flowers V (rflowers5.netlify.com).
}
\author{
\strong{Maintainer}: Pedro J. Perez \email{pedro.j.perez@example.com}

}
\keyword{internal}
